Create Mapping Util File
